Refactored format method to stringUtils #120 #118
